refactor: separate between model and metadata
