+++ tests/integration/model/helpers/requestSender.ts
@@ -1,0 +1,10 @@
+import * as supertest from 'supertest';
+import { IngestionPayload } from '../../../../src/common/interfaces';
+
+export class ModelRequestSender {
+  public constructor(private readonly app: Express.Application) {}
+
+  public async createModel(payload: IngestionPayload): Promise<supertest.Response> {
+    return supertest.agent(this.app).post('/model/ingestion').set('Content-Type', 'application/json').send(payload);
+  }
+}
