+++ tests/integration/model/modelController.spec.ts
@@ -3,7 +3,7 @@ import { trace } from '@opentelemetry/api';
 import { StatusCodes } from 'http-status-codes';
 import { ProductType } from '@map-colonies/mc-model-types';
 import mockAxios from 'jest-mock-axios';
-import { randFutureDate, randNumber, randPastDate, randSentence, randWord } from '@ngneat/falso';
+import { randFutureDate, randNumber, randPastDate, randWord } from '@ngneat/falso';
 import { ILookupOption } from '../../../src/externalServices/lookupTables/interfaces';
 import {
   createMetadata,
@@ -15,17 +15,14 @@ import {
   createWrongFootprintCoordinates,
   createWrongFootprintSchema,
   getBasePath,
-  createUuid,
-  createUpdatePayload,
-  createUpdateStatusPayload,
 } from '../../helpers/helpers';
 import { getApp } from '../../../src/app';
 import { SERVICES } from '../../../src/common/constants';
 import { IngestionPayload } from '../../../src/common/interfaces';
-import { MiddlewareRequestSender } from './helpers/requestSender';
+import { ModelRequestSender } from './helpers/requestSender';
 
-describe('MiddlewareController', function () {
-  let requestSender: MiddlewareRequestSender;
+describe('ModelController', function () {
+  let requestSender: ModelRequestSender;
   beforeEach(function () {
     const app = getApp({
       override: [
@@ -34,14 +31,14 @@ describe('MiddlewareController', function () {
       ],
       useChild: true,
     });
-    requestSender = new MiddlewareRequestSender(app);
+    requestSender = new ModelRequestSender(app);
   });
 
   afterEach(function () {
     mockAxios.reset();
   });
 
-  describe('POST /ingestion', function () {
+  describe('POST /model/ingestion', function () {
     describe('Happy Path 🙂', function () {
       describe('Sphere', function () {
         it('should return 201 status code and the added model', async function () {
@@ -378,165 +375,6 @@ describe('MiddlewareController', function () {
 
         expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
         expect(response.body).toHaveProperty('message', 'there is a problem with catalog');
-      });
-    });
-  });
-
-  describe('POST /update/{identifier}', function () {
-    describe('Happy Path 🙂', function () {
-      it(`should return 200 status code and metadata if payload is valid`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        const expected = randSentence();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.OK });
-        mockAxios.get.mockResolvedValueOnce({ data: [{ value: payload.classification }] as ILookupOption[] });
-        mockAxios.patch.mockResolvedValueOnce({ status: StatusCodes.OK, data: expected });
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.OK);
-        expect(response.body).toBe(expected);
-      });
-    });
-
-    describe('Bad Path 😡', function () {
-      it(`should return 400 status code if classification is not valid`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        const classification = randWord();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.OK });
-        mockAxios.get.mockResolvedValueOnce({ data: [{ value: classification }] as ILookupOption[] });
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.BAD_REQUEST);
-        expect(response.body).toHaveProperty('message', `classification is not a valid value.. Optional values: ${classification}`);
-      });
-
-      it(`should return 400 status code if record does not exist in catalog`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.NOT_FOUND });
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.BAD_REQUEST);
-        expect(response.body).toHaveProperty('message', `Record with identifier: ${identifier} doesn't exist!`);
-      });
-    });
-
-    describe('Sad Path 😥', function () {
-      it(`should return 500 status code if during validation, catalog didn't return as expected`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.CONFLICT });
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'Problem with the catalog during validation of record existence');
-      });
-
-      it(`should return 500 status code if lookup-tables is not working properly`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.OK });
-        mockAxios.get.mockRejectedValueOnce(new Error('lookup-tables error'));
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'there is a problem with lookup-tables');
-      });
-
-      it(`should return 500 status code if catalog is not working properly`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        mockAxios.get.mockRejectedValueOnce(new Error('catalog error'));
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'there is a problem with catalog');
-      });
-
-      it(`should return 500 status code if during sending request, catalog didn't return as expected`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdatePayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.OK });
-        mockAxios.get.mockResolvedValueOnce({ data: [{ value: payload.classification }] as ILookupOption[] });
-        mockAxios.patch.mockResolvedValueOnce({ status: StatusCodes.CONFLICT });
-
-        const response = await requestSender.updateMetadata(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'there is an error with catalog');
-      });
-    });
-  });
-
-  describe('POST /update/status/{identifier}', function () {
-    describe('Happy Path 🙂', function () {
-      it(`should return 200 status code and metadata if payload is valid`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdateStatusPayload();
-        const expected = randSentence();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.OK });
-        mockAxios.patch.mockResolvedValueOnce({ status: StatusCodes.OK, data: expected });
-
-        const response = await requestSender.updateStatus(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.OK);
-        expect(response.body).toBe(expected);
-      });
-    });
-
-    describe('Bad Path 😡', function () {
-      it(`should return 400 status code if record does not exist in catalog`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdateStatusPayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.NOT_FOUND });
-
-        const response = await requestSender.updateStatus(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.BAD_REQUEST);
-        expect(response.body).toHaveProperty('message', `Record with identifier: ${identifier} doesn't exist!`);
-      });
-    });
-
-    describe('Sad Path 😥', function () {
-      it(`should return 500 status code if during validation, catalog didn't return as expected`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdateStatusPayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.CONFLICT });
-
-        const response = await requestSender.updateStatus(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'Problem with the catalog during validation of record existence');
-      });
-
-      it(`should return 500 status code if catalog is not working properly`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdateStatusPayload();
-        mockAxios.get.mockRejectedValueOnce(new Error('catalog error'));
-
-        const response = await requestSender.updateStatus(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'there is a problem with catalog');
-      });
-
-      it(`should return 500 status code if during sending request, catalog didn't return as expected`, async function () {
-        const identifier = createUuid();
-        const payload = createUpdateStatusPayload();
-        mockAxios.get.mockResolvedValueOnce({ status: StatusCodes.OK });
-        mockAxios.patch.mockResolvedValueOnce({ status: StatusCodes.CONFLICT });
-
-        const response = await requestSender.updateStatus(identifier, payload);
-
-        expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-        expect(response.body).toHaveProperty('message', 'there is an error with catalog');
       });
     });
   });
